feat: replace disable directives comments
